Reorganized database related files
